+++ app/templates/views/service-settings.html
@@ -204,7 +204,7 @@
           </li>
           <li class="bottom-gutter">
           {% if can_receive_inbound %}
-            <a href="{{ url_for('.service_set_sms_sender', service_id=current_service.id, set_inbound_sms=True) }}" class="button">
+            <a href="{{ url_for('.service_set_inbound_number', service_id=current_service.id, set_inbound_sms=False) }}" class="button">
               Stop inbound sms
             </a>
           {% else %}
